+++ core/src/hal/thermal.rs
@@ -4,18 +4,18 @@ use crate::hal::monitored_file::MonitoredFile;
 
 pub struct ThermalSensor {
     monitor: Option<MonitoredFile<16>>,
-    default_val: f64,
+    default_val: f32,
 }
 
 impl ThermalSensor {
-    pub fn new(path: &str, default_val: f64) -> Self {
+    pub fn new(path: &str, default_val: f32) -> Self {
         let monitor = MonitoredFile::new(path).ok();
         Self {
             monitor,
             default_val,
         }
     }
-    pub fn read(&mut self) -> f64 {
+    pub fn read(&mut self) -> f32 {
         let monitor = match self.monitor.as_mut() {
             Some(m) => m,
             None => return self.default_val,
@@ -23,7 +23,7 @@ impl ThermalSensor {
         match monitor.read_value() {
             Ok(content) => {
                 let s = content.trim();
-                match s.parse::<f64>() {
+                match s.parse::<f32>() {
                     Ok(val) => {
                         let abs = val.abs();
                         if abs >= 10_000.0 {
